+++ R/settings.R
@@ -23,7 +23,13 @@ renv_settings_validate <- function(name, value) {
 
   # otherwise, validate the user-provided value
   validate <- `_renv_settings`[[name]]$validate
-  if (validate(value))
+  ok <- case(
+    is.character(validate) ~ value %in% validate,
+    is.function(validate)  ~ validate(value),
+    TRUE
+  )
+
+  if (identical(ok, TRUE))
     return(value)
 
   # validation failed; warn the user and use default
@@ -303,8 +309,8 @@ renv_settings_impl <- function(name, validate, default, update) {
 #' For example:
 #'
 #' \preformatted{
-#' options(renv.settings = list(snapshot.type = "simple"))
-#' options(renv.settings.snapshot.type = "simple")
+#' options(renv.settings = list(snapshot.type = "all"))
+#' options(renv.settings.snapshot.type = "all")
 #' }
 #'
 #' If both of the `renv.settings` and `renv.settings.<name>` options are set
@@ -350,8 +356,8 @@ settings <- list(
 
   snapshot.type = renv_settings_impl(
     name     = "snapshot.type",
-    validate = function(x) x %in% c("custom", "simple", "packrat"),
-    default  = "packrat",
+    validate = c("all", "custom", "implicit", "explicit", "packrat", "simple"),
+    default  = "implicit",
     update   = NULL
   ),